cleanup: move tutorial view to seperate component
